Add PUBLISH_ARTIFACTS parameter to pipeline builds
We just turned on daily master builds for this project, but we don't want every
master build to publish new artifacts. To ensure this only happens when we
intend for it to, we have added a build parameter that is by default false but
can be turned on when we need to publish new AMIs / CFTs.